fixed: [OS X] some compile issues
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@12413 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/guilib/GUIListItem.cpp
@@ -173,7 +173,7 @@ void CGUIListItem::Serialize(CArchive &ar)
     ar << m_strIcon;
     ar << m_bSelected;
     ar << m_overlayIcon;
-    ar << m_mapProperties.size();
+    ar << (int)m_mapProperties.size();
     for (std::map<CStdString, CStdString, icompare>::const_iterator it = m_mapProperties.begin(); it != m_mapProperties.end(); it++)
     {
       ar << it->first;
